replace nanomsg with msgqueue, fix trigger:poll() semantics
nanomsg is too big. the only place where we used it (in core) was to
deliver events from C threads into the Lua scheduler's event queue.

now that we have `msgqueue`, nanomsg can be moved out of core into a
separate package.

about the poll semantics of trigger:

POSIX variants of poll (select, poll, epoll) check if a set of file
descriptors are readable/writable but do not execute a read or write
on the file descriptors - this is the responsibility of the caller

I modified trigger:poll() to follow these semantics and provided a
separate trigger:read() which can be used to read the value after a
successful poll. the read() call resets (or decrements) the internal
counter of the trigger.

when you need a poll followed by a read, you can use trigger:wait() as
a shortcut

msgqueue also gained wait() and reset_trigger() methods which are
proxied to trig_r:wait() and trig_r:read() respectively